Workaround fix for RMAN-21063.
Make xpu_slow_mode the default again. I'm not
sure why the callback method is getting stuck when
there's a dispaly edit; gdb shows scene graph is
stuck on trying to acquire a mutex during an edit.
Gut feeling is maybe this has something to do with
the python GIL.