Fixes for GitHub #689 and GitHub #694
* Check the length of the socket links before accessing.
Eventhough "is_linked" is True, the link might not exists
* Change new_links to be a class variable instead of an instance property in RendermanShadingNode class.
We now have the update method in the base class do the looping over new_links, instead of the child classes.
I believe we were crashing because both
the to and from nodes were trying to check if the link is valid.
* We add a new method called accept_link on the child classes of RendermanShadingNode.
Returning true from this method tell us that the link is acceptable. We only call accept_link on the to node, and not the from node.