Fix issue reported on Discrod. The "attribute" parameter was not showing up in PxrCryptomatte.
"attribute" has conditional vis op that looks at the "layer" parameter and see if it's set to "".
Unfortunately, Blender does not allow empty strings to be used in their EnumProperty so we substitute those fo "__empty__".
To fix this specfic issue we modify conditional expression to also check for "__empty__", when the value is set to "".

Also, fix another issue where disconnecting nodes in the shader graph was not triggering an IPR update.