Add support for Blender 4.x
* Biggest change is the NodeSocketInterface. We seem to be
unable to dynamically create the classes for our NodeSockets, so
for now, just explicitly create them.
* There are changes in the mesh API. There's no longer face maps
so remove support for using that as holes for subdivs. Also,
the methods for getting smooth normals and edg creases have
changed.
* A couple of strings/enums in the gpu module have changed
* Fix a couple of other bugs unrelated to 4.0
* Unfortunately, there's a bug in Blender that causes it to crash
when using our shading nodes. The BF have fixed it, but only
in their alpha 4.1 branch, so we can only support Blender 4.1.