Fix an issue reported by Eugene.
We would ge a traceback when creating a node from the search
menu.
Add some error checking  in our create_node operator. We might
not actually have a node_tree instance. If we don't, check for it
in the space_area context.
Also, add an operator zip up the addon. This might be useful if
we made changes to the code, and want to share with a user for them to test.