Fix for CUSP-8148
When we import a preset that includes nodes with ramps in them,
we need to remove all of the knots from the ramp first. The ramp
from the preset might have lesser knots than the default, and this
was the case for the stylized presets in the test case.

Also, make sure we don't double the knots when we export a preset.
The doubling of the knots should be done by the DCC that's
importing the preset.