Fix for CUSP-8201
The previous fix for soloing did not work in this case,
because the shader nodes are out of order.
Instead, expand gather_nodes function to deal with
solo nodes that are inside of node groups. If we are in
this situation, find which nodetree the group node is in
and then gather nodes from that nodetree.

Also, fix an issue where we were not handling sockets
to NodeGroupInput nodes with the same names.